// old/data/type.ts
export interface ITEM_STACK_DATA {
	/**
	 * TODO: 物品类型
	 */
	type: string;
	/**
	 * TODO: 物品数量
	 */
	amount?: number;
	/**
	 * TODO: 物品名称
	 */
	name?: string;
	/**
	 * TODO: 物品词缀
	 */
	lore?: string[];
	/**
	 * TODO: 物品动态属性
	 */
	property?: Map<string, boolean | number | string | server.Vector3>;
};